+++ client/src/test/java/replicant/ReplicantRuntimeTest.java
@@ -328,7 +328,7 @@ public class ReplicantRuntimeTest
 
     entry1.getRateLimiter().fillBucket();
     runtime.deactivate();
-    verify( service1.getTransport(), times( 1 ) ).disconnect();
+    verify( service1.getTransport(), times( 1 ) ).requestDisconnect();
 
     reset( service1.getTransport() );
 
@@ -338,13 +338,13 @@ public class ReplicantRuntimeTest
     safeAction( () -> service1.setState( ConnectorState.CONNECTING ) );
     entry1.getRateLimiter().fillBucket();
     runtime.deactivate();
-    verify( service1.getTransport(), never() ).disconnect();
+    verify( service1.getTransport(), never() ).requestDisconnect();
 
     runtime.activate();
     safeAction( () -> service1.setState( ConnectorState.DISCONNECTING ) );
     entry1.getRateLimiter().fillBucket();
     runtime.deactivate();
-    verify( service1.getTransport(), never() ).disconnect();
+    verify( service1.getTransport(), never() ).requestDisconnect();
 
     // set service state to DISCONNECTED so no action required
 
@@ -352,7 +352,7 @@ public class ReplicantRuntimeTest
     safeAction( () -> service1.setState( ConnectorState.DISCONNECTED ) );
     entry1.getRateLimiter().fillBucket();
     runtime.deactivate();
-    verify( service1.getTransport(), never() ).disconnect();
+    verify( service1.getTransport(), never() ).requestDisconnect();
 
     // set service state to ERROR so no action required
 
@@ -360,7 +360,7 @@ public class ReplicantRuntimeTest
     safeAction( () -> service1.setState( ConnectorState.ERROR ) );
     entry1.getRateLimiter().fillBucket();
     runtime.deactivate();
-    verify( service1.getTransport(), never() ).disconnect();
+    verify( service1.getTransport(), never() ).requestDisconnect();
 
     // set service state to connected but rate limit it
 
@@ -368,7 +368,7 @@ public class ReplicantRuntimeTest
     safeAction( () -> service1.setState( ConnectorState.CONNECTED ) );
     entry1.getRateLimiter().setTokenCount( 0 );
     runtime.deactivate();
-    verify( service1.getTransport(), never() ).disconnect();
+    verify( service1.getTransport(), never() ).requestDisconnect();
 
     // set service state to connected
 
@@ -376,7 +376,7 @@ public class ReplicantRuntimeTest
     safeAction( () -> service1.setState( ConnectorState.CONNECTED ) );
     entry1.getRateLimiter().fillBucket();
     runtime.deactivate();
-    verify( service1.getTransport(), times( 1 ) ).disconnect();
+    verify( service1.getTransport(), times( 1 ) ).requestDisconnect();
   }
 
   @Test
@@ -406,8 +406,8 @@ public class ReplicantRuntimeTest
     entry1.getRateLimiter().fillBucket();
     entry3.getRateLimiter().fillBucket();
     runtime.deactivate();
-    verify( service1.getTransport(), times( 1 ) ).disconnect();
-    verify( service3.getTransport(), times( 1 ) ).disconnect();
+    verify( service1.getTransport(), times( 1 ) ).requestDisconnect();
+    verify( service3.getTransport(), times( 1 ) ).requestDisconnect();
   }
 
   @Test